Move filterController files
