+++ app/filter/filtercontroller.cpp
